add amp velocity and click noise

+++ include/output_limiter.h
@@ -4,8 +4,8 @@
 
 namespace OutputLimiter {
 
-// Applies master gain and brickwall limiting to one mono block.
+// Applies master gain and brickwall limiting to one wide-sample mono block.
 // The returned buffer remains valid until the next call.
-const int16_t* process(const int16_t* inputSamples);
+const int16_t* process(const int32_t* inputSamples);
 
 }  // namespace OutputLimiter
